Add ava, remove old tests
