+++ package.json
@@ -7,7 +7,7 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "ava"
   },
   "repository": {
     "type": "git",
@@ -18,5 +18,9 @@
   "bugs": {
     "url": "https://github.com/ProtonMail/sieve.js/issues"
   },
-  "homepage": "https://github.com/ProtonMail/sieve.js#readme"
+  "homepage": "https://github.com/ProtonMail/sieve.js#readme",
+  "dependencies": {},
+  "devDependencies": {
+    "ava": "1.0.0-beta.4"
+  }
 }
